Refactor Program.cs
* Extract all possible logic into a new RunnerService.cs for unit testing.
